move the follow-up requests test in the external tests suite
